+++ tests/functional_tests/ping_pong_tester.cpp
@@ -77,12 +77,12 @@ PingPongTester::PingPongTester(TesterArguments args) : Tester(args) {
 
 PingPongTester::~PingPongTester() { rocshmem_free(r_buf); }
 
-void PingPongTester::resetBuffers(uint64_t size) {
+void PingPongTester::resetBuffers(size_t size) {
   memset(r_buf, 0, sizeof(int) * args.num_wgs);
 }
 
 void PingPongTester::launchKernel(dim3 gridSize, dim3 blockSize, int loop,
-                                  uint64_t size) {
+                                  size_t size) {
   size_t shared_bytes = 0;
 
   hipLaunchKernelGGL(PingPongTest, gridSize, blockSize, shared_bytes, stream,
@@ -93,4 +93,4 @@ void PingPongTester::launchKernel(dim3 gridSize, dim3 blockSize, int loop,
   num_timed_msgs = loop;
 }
 
-void PingPongTester::verifyResults(uint64_t size) {}
+void PingPongTester::verifyResults(size_t size) {}
