+++ tests/functional_tests/ping_pong_tester.hpp
@@ -42,12 +42,12 @@ class PingPongTester : public Tester {
   virtual ~PingPongTester();
 
  protected:
-  virtual void resetBuffers(uint64_t size) override;
+  virtual void resetBuffers(size_t size) override;
 
   virtual void launchKernel(dim3 gridSize, dim3 blockSize, int loop,
-                            uint64_t size) override;
+                            size_t size) override;
 
-  virtual void verifyResults(uint64_t size) override;
+  virtual void verifyResults(size_t size) override;
 
   int *r_buf;
 };
