+++ tests/functional_tests/signaling_operations_tester.hpp
@@ -37,12 +37,12 @@ class SignalingOperationsTester : public Tester {
   virtual ~SignalingOperationsTester();
 
  protected:
-  virtual void resetBuffers(uint64_t size) override;
+  virtual void resetBuffers(size_t size) override;
 
   virtual void launchKernel(dim3 gridSize, dim3 blockSize, int loop,
-                            uint64_t size) override;
+                            size_t size) override;
 
-  virtual void verifyResults(uint64_t size) override;
+  virtual void verifyResults(size_t size) override;
 
   int sig_op;
   char *s_buf = nullptr;
